Move ProcessFinishedStatus to core module
