+++ java/core/src/main/java/sleeper/core/status/ProcessFinishedStatus.java
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.job.status;
+package sleeper.core.status;
 
 import sleeper.core.record.process.RecordsProcessedSummary;
-import sleeper.core.status.ProcessStatusUpdate;
 
 import java.time.Instant;
 import java.util.Objects;
